# api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Flight Data API",
    "description": "API for accessing and analyzing flight data",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    }
  },
  "paths": {
    "/flights": {
      "get": {
        "tags": ["Flights"],
        "summary": "Get flight data with filtering and pagination",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Filter flights from this date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Filter flights until this date (YYYY-MM-DD)"
          },
          {
            "name": "origin",
            "in": "query",
            "type": "string",
            "description": "Filter by origin airport code"
          },
          {
            "name": "destination",
            "in": "query",
            "type": "string",
            "description": "Filter by destination airport code"
          },
          {
            "name": "cursor",
            "in": "query",
            "type": "string",
            "description": "Pagination cursor from previous response"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 100,
            "description": "Number of records to return (max 1000)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "flights": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "next_cursor": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                },
                "total": {
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/metrics": {
      "get": {
        "tags": ["Metrics"],
        "summary": "Get flight metrics with date filtering",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Calculate metrics from this date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Calculate metrics until this date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "total_flights": {
                  "type": "integer"
                },
                "delay_rate": {
                  "type": "number"
                },
                "cancellation_rate": {
                  "type": "number"
                },
                "diversion_rate": {
                  "type": "number"
                },
                "avg_departure_delay": {
                  "type": "number"
                },
                "avg_arrival_delay": {
                  "type": "number"
                },
                "max_departure_delay": {
                  "type": "number"
                },
                "max_arrival_delay": {
                  "type": "number"
                },
                "top_routes": {
                  "type": "object"
                },
                "top_carriers": {
                  "type": "object"
                },
                "date_range": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/airports": {
      "get": {
        "tags": ["Airports"],
        "summary": "Get list of all airports",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "airports": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "count": {
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "BasicAuth": []
          }
        ]
      }
    }
  }
} 